Make PipelineObserver internal
There's no need for this to be exposed to the wider world. Move it to
its own file at the same time.
