+++ src/Waives.Reactive/Pipeline.cs
@@ -171,29 +171,4 @@ namespace Waives.Reactive
             return observer.SubscribeTo(_pipeline);
         }
     }
-
-    public class PipelineObserver : IObserver<WaivesDocument>
-    {
-        private readonly Action _onPipelineCompleted;
-
-        internal PipelineObserver(Action onPipelineCompleted)
-        {
-            _onPipelineCompleted = onPipelineCompleted;
-        }
-
-        public void OnCompleted()
-        {
-            _onPipelineCompleted();
-        }
-
-        public void OnError(Exception error)
-        {
-
-        }
-
-        public void OnNext(WaivesDocument value)
-        {
-
-        }
-    }
 }
